Removed broken "log to file" feature
While implementing "log to STDOUT" I realised that "log to file" was actually
broken to begin with, and even though it creates/touches a finatra.log file it
doesn't really write to it.

If no one needs this feature (and given no one complained one can assume
so...) it's better to kill it.