finatra/inject: singletonStartup/Shutdown methods
Problem:

To invoke startup functions within a Module, you have to call
singletonStartup within a configure().  singletonStartup mutates
the callback list of the module. This makes the module mutable, and
it can't be a singleton because when Guice calls its configure()
method, you mutate global state. The problem is exacerbated because
modules that compose it also can't be singletons even if the top
level module doesn't use singletonStartup. Thus it's not always
obvious when a module can be a singleton or not.

tl;dr global mutable state is bad

Solution:

Instead of appending callbacks to a list, use a method override.

override def singletonStartup(injector: Inject): Unit

Result:

Calling configure() will never mutate any state. Modules are more
side-effect free and can be objects.

RB_ID=685180