- fix method validation error rendering
Problem:

We use a string to represent the field (e.g. "id").  For nested fields, the idea is to prefix the string (e.g. "foo.id"). Since method validations are applied to the object itself and not a particular field within that object, we incorrectly represent the field part of CaseClassValidationException with the empty string. This breaks nested fields because calling .nest() will always prefix the field with the parent field + ".". So you get weird errors like "params.: some error message".

Solution:

Represent fields in CaseClassValidationException as Seq[String], which more closely reflects the semantics of what we're trying to represent (a path that identifies a field). This solves the problem because we simply call `pathToField.mkString(".")` when we need the field, which is more resilient against corner cases like the one described above.

Result:

Error messages are correctly displayed.

RB_ID=719546