inject-server: Track and provide finagle-metrics gauge for finagle version being used
Problem

When running your service in a large cluster it would useful to know if all your
shards are running against the same version of finagle as this can be important.
An easy way to track that would be to graph a metric that is provided by your
server that gives you a discreet representation of the finagle version.
Currently, in Finatra we do not provide this.

Solution

Add logic to capture and stat as a gauge a representation of the finagle version
the current Finatra server is running against.

Result

This stat can be graphed per shard for easy detection of outliers in a cluster.

RB_ID=721898