twitter-server-internal: Remove dependency on finagle-http
finagle-http will soon be deprecated in favor of finagle-httpx. In
preparation for this change, we're assisting each team in converting
their targets that depend on finagle-http to finagle-httpx.

This RB removes the finagle-http dependency from twitter-server-internal
and adds it as a direct depedency to projects that were using it
transitively. We'll follow up rbs that convert each project
individually. However, If you'd like to help out by converting your
project before we get to it, that would be much appreciated :)

This spreadsheet tracks what targets have already been converted:
https://docs.google.com/spreadsheets/d/1ABj_r642UUjN-zNN9PM8AS_2cC06lSmks9SyGA20OGY/edit#gid=473845749.

Background: go/httpx

RB_ID=741454
TBR=true