http - Add support for "uncached" mustache templates.
Problem:
Currently, Finatra does not allow for mustache templates to be reloaded on every call to render a template meaning that you must recompile to reload template changes. This is especially burdensome in local development.

Resolution:
Per this issue: https://github.com/spullara/mustache.java/issues/117 the recommended way to provide this functionality is to create a new MustacheFactory on every call to compile the template thus implicitly getting a clean cache every time. The MustacheModule will now provide this functionality piggy-backing on the setting of the `local.doc.root` flag for the FileResolver. If this flag is set, templates will be reloaded on every render. Note, that the interplay between the `mustache.temaplates.dir` and the `local.doc.root` flag is as follows:
  - in "local file mode" (e.g., when `local.doc.root` is set to a **non-empty** value) we will try to load a template first from the absolute path under `mustache.templates.dir`
  - if it is not found it will then be loaded from a location of `mustache.templates.dir` relative to `local.doc.root`, e.g, `local.doc.root`/`mustache.templates.dir`/template.mustache.

Additionally, we've added `MustacheService#createString` which allows you to programmatically render a template to a string for embedding inside of a field in a response. For this we have also moved the `MustacheService` from an internal package.

Result:
For local development a developer can now choose to configure filesystem reloading of mustache templates. And also now has the ability to render templates to strings programmatically.

RB_ID=746863
TBR=true