Add exception mapper case for c.t.f.Failure.
Problem:
The FinatraDefaultExceptionMapper does not handle c.t.f.Failures.  This means that a c.t.f.Failure that wraps an exception that FinatraDefaultExceptionMapper can map will result in a 500 instead of the appropriate Response.

Solution:
Add a case to FinatraDefaultExceptionMapper for unwrapping c.t.f.Failures up to a maximum depth.

Result:
FinatraDefaultExceptionMapper will correctly map c.t.f.Failures that wrap an exception that it knows how to map.

RB_ID=759408