finatra - change EmbeddedTwitterServer to accept a base com.twitter.server.TwitterServer
Problem
The EmbeddedTwitterServer does not properly accept a base com.twitter.server.TwitterServer.
This is an unncessary limitation of the current implementation and does not mirror
EmbeddedApp which does accept a base com.twitter.app.App.

Resolution
Change the EmbeddedTwitterServer to accept a com.twitter.server.TwitterServer and address
the necessary code changes as a result of the change. We also a address the issue of not
waiting for the EmebeddedApp future on close which was masking some failures as the errors
happened in a seperate orphaned thread and were thus not seen.

Result
Any base com.twitter.server.TwitterServer can be used with the EmbeddedTwitter.

RB_ID=760648
TBR=true