http - remove Ports#httpExternalSocketAddress
Problem
Exposing a pure SocketAddress type is effectively useless as one would have to cast
it to a meaningful type, e.g., InetSocketAddress. Since this field was exposed early
without any due dilligence in regard to utility, we've since decided to remove it to
simplify the codebase.

Resolution
Removal of Ports#httpExternalSocketAddress

Result
Less lines of useless code.

RB_ID=766116