finatra: Change startupFailedThrowable to volatile var in EmbeddedApp
Problem

Within EmbeddedApp, startupFailedThrowable is used to track if an
exception was thrown. This value is written and read on two
different threads. It is possible that updates made in one thread
will not be seen by the other thread. This could cause exceptions
to not be propagated properly.

Solution

Make startupFailedThrowable a volatile var so that if an exception
occurs on startup that it should be immediately seen within
waitForAppStarted.

Result

Exceptions that occur within the mainRunnerFuturePool are always
seen by waitForAppStarted.

RB_ID=766576