http - Better error messaging for "empty" route callbacks in Controllers
Problem
Currently, it is possible to define a route in a Controller with an "empty"
callback, e.g., a callback with no parameter if the callback returns a String.
Scala infers this as a function of Int => Char which is the function signature
for StringOps.apply. If you do this you'll end up in the last case in the
CallbackConverter#createRequestCallback which will end up delegating to Jackson
which tries to find a Jackson deserializer that goes from a Request to an Int.
This fails with an unhelpful error message.

Resolution
Since it is highly unlikely that someone really wants to use a callback that
takes in an Int parameter we've added logic to catch when the parameter type
of the callback is an Int and now return a more useful error message.
Unfortunately, this is only at runtime but ideally will get caught in testing.

Result
More helpful error messaging when creating incorrectly parameterized routes in
Controllers.

RB_ID=774029