thrift - Add flag and functionality for announcing Thrift Server
Problem

We don't currently expose the ability to announce a Thrift server
akin to the functionality in the HttpServer (via BaseHttpServer).

Resolution

Add the 'thrift.announce' flag and functionality to invoke
server.announce when the flag is set to bring functionality on
par with Http support.

Result

Users will be able to optionally announce their Thrift server by
setting the 'thrift.announce' flag appropriately.

RB_ID=787543