http - Ensure RouteInfo is available for use in filters Problem
Currently routing happens after filters which means that filters cannot
make use of the RouteInfo seeded into the Request.ctx. This would be
problematic for any filter that wanted to expose information for the
given route requested, e.g., an authentication filter providing that it
failed authentication to a specific route.

Resolution

Changed the HttpRouter to apply routing first before filtering.

Result

Routing information, e.g., the seeded RouteInfo object is available in
the Request.ctx for filters to use.

RB_ID=786561
TBR=true