finatra: Fix Race Condition in Exception Startup Integration Test
Problem

An integration test within StartupIntegrationTest is
non-deterministic to where the Exception bubbling up
and appStarted are competing to see who finishes first
and is seen by the loop which is checking both values.

Solution

For this test, add a method to make sure appStarted is
not set at all.

Result

As long as the exception is thrown within the test timeout
period, the test should always pass now.

RB_ID=769699