finatra-http - Fix handling of failed JSON parsing in HttpClient
Problem

In the case that HttpClient#executeJson fails to parse the
successful response body into an instance of the type T
we do not properly handle the parsing failure.

Resolution

Change the code to properly handle the parsing failure by
wrapping the call in a Try and pattern-matching on the
result which branches into either a properly parsed
response body or a failed future wrapping an
HttpClientException which encodes the underlying exception.

Result

Better handling when parsing the response body fails even
on a successful response.

RB_ID=791067