Pass disableTestLogging argument to EmbeddedApp
Problem
The disableTestLogging constructor parameter in EmbeddedHttpServer,
EmbeddedThriftServer and EmbeddedTwitterServer is not passed to the EmbeddedApp
constructor. This means the parameter does not have any effect and all tests
produce a full verbose output even when this is not desired.

Solution
Pass the disableTestLogging argument to the parent constructors.

Signed-off-by: Ryan O'Neill <ryano@twitter.com>

RB_ID=793017