finatra-thrift - Improve Finatra thrift controller implementation
Problem

The current implementation of the Thrift controller in Finatra
suffers from the need to special-case logic for methods with
greater than 22 args.

Resolution

With a recent scrooge change we are able to simplify the thrift
controller to always handle a Function1 from
ThriftMethod.Args => ThriftMethod.Result. This allows us to
remove the special-case code for methods greater than 22 args
and also more closely mirros the Http controller route DSL
in which the callback is a Function1 of Request => Any.

Result

Simpler code for Thrift controllers.

RB_ID=793713