finatra-jackson - Improve JsonDiff to return diff in Exception message.
Problem

JsonDiff currently only logs when a difference is encountered. The diff
is not included in the message of the thrown Exception which means that
it requires developers to scrape the stdout for the cause of the test
failure.

Resolution

Include the Json diff output in the Exception message of the thrown
Exception in the JsonDiff test class. Introduce a production line
utility, JsonDiffUtil with the primary responsibility of
calculating Json diffs. In the case where a difference is found a
Some(JsonDiffResult) is returned otherwise a None. Change JsonDiff
to use this new core utility.

Result

Developers will have the ability to inspect the Exception message to
understand the reason a test failed and the JsonDiffUtil will be
usable without a test-jar dependency.

RB_ID=794318