finatra - Bring parity to ThriftServer
Problem

We don't support for disabling of the TwitterServer Http
Admin Interface in the ThriftServer akin to HttpServer.
Additionally, we don't allow for overriding of default values
for flags in ThriftServer, nor for admin announcing in a
ThriftServer, or to override if the server should fail fast
on flag parsing (true by default). We also did not have a test
case covering a "combined" http and thrift server.

Resolution

Allow for programmatic overriding of default flag values defined
in ThriftServer. Move common functionality for
HttpServer/ThriftServer to TwitterServer so that both server
types can choose to disable the Http Admin Interface.

Result

Better feature parity between HttpServer and ThriftServer.

RB_ID=797365
TBR=true