finatra-thrift - Support setting thrift server label
Problem

We don't expose a way for a thrift server to configure
it's label with Finagle. Additionally, we are adding the
flag "thrift.port" multiple times when starting an
EmbeddedThriftServer b/c it is manually added in the
constructor of EmbeddedThriftServer, then we override
combineArgs in EmbeddedThriftServer and we then also
override combineArgs in ThriftClient making it such that
and EmbeddedThriftServer adds the flag 3 times.

Resolution

Change the ThriftServer#postWarmup logic to configure
the ThriftMux.server withLabel. Create a new
ThriftServer#thriftServerNameFlag which mirrors
HttpServer#httpServerNameFlag and is overridable in
code. This flag value is then used when creating the
thrift server with Finagle as the label.

Clean up adding of the "thrift.port" flag w.r.t to
EmbeddedThriftServer and ThriftClient.

Result

More configurable thrift servers. Better
EmbeddedThriftServer for testing.

RB_ID=798284
TBR=true