finatra-thrift: Fix type agnostic filtering
Problem

The current implementation of the TypeAgnostic filters in Finatra suffers from
the drawback of creating a new instance of a filter per request. This has been
noticed to cause high ParNew GC under load.

Resolution

Unify all the things. With the TypeAgnostic filters as a catalyst work to
better unify the filtering API across both the HttpRouter and ThriftRouter.
Remove deprecated functionality from the ThriftRouter and package classes in
finatra-thrift accordingly to be consistent with like classes in finatra-http.
Simplify the usage of the Slf4jBridgeModule for MDC and sl4j bridges
initialization.
  - The "Slf4jBridgeModule" is now a framework module and now added by default
    if you extend HttpServer or ThriftServer.
  - Remove deprecated functionality in the ThriftRouter
  - Create corresponding ThriftFilters for LoggingMDCFilter and TraceIdMDCFilter
  - ensure consistent "javax.inject" imports for "Inject" and "Singleton"

Result

Finatra thrift functionality should now be on par with http and ready to be
fully documented and released.

RB_ID=802947
TBR=true