finatra - Dynamically determine if Slf4jBridgeHandler should be installed
Problem

If you are using slf4j-jdk14 as a logging implementation we
currently add the jul-to-slf4j bridge transitively which is necessary
as we add the Slf4jBridgeModule as a framework module by default.
This dependency chain results in an infinite loop.

Resolution

Dynamically determine if the Slf4jBridgeModule should install the
Slf4jBridgeHandler based on if the JDK14LoggerFactory is present
on the classpath. This class comes from the slf4j-jdk14 dependency.

Result

The ability to use the slf4j-jdk14 logging implementation.

RB_ID=804190