finatra/http: Refactor Template Lookup to New Class
Problem

Looking up which mustache file is associated with a particular
class is currently an implementation detail of
MustacheMessageBodyWriter. It would be beneficial to separate
out the logic and be able to reuse it internally.

Solution

Refactor MustacheMessageBodyWriter to use a
MustacheTemplateNameLookup instance where the lookup and caching
logic now reside.

RB_ID=805317