finatra-http: Update admin UI indexing
Problem

Adding routes to the admin UI index is a bit too aggressive and
not fully-backwards compatible which causes for unexpected
behavior.

Solution

Update routing to ensure it's an opt-in for adding routes to
the admin UI index.

Any constant admin route will be added to the admin index.
Only constant GET routes are eligible to be added to the UI.
These routes may start with only /admin.

If you want the Finatra router (instead of the TwitterServer
admin router) to be invoked the route MUST start with
/admin/finatra. These routes are also not eligible to be
added to the UI index.

Result

Better admin routing.

RB_ID=809338