finatra-jackson: Support Min/Max/Range validations on Double/Float fields
Problem

Currently, if a java.lang.Number type is validated with these validations
it is cast to the long value to do the validation. This can cause
unexpected behavior if you are using a floating point value in your case
class (Double for instance) If you have a @Min(0) val: Double and the
incoming value is -0.5, the validation passes because when -0.5 is turned
into a long value it is truncated towards zero.

Solution

Support double values in Min/Max/Range validators

Result

Better functionality for validators and less confusing behavior.

RB_ID=809821