finatra-thrift: Support for using TLS with thrift server
Problem

It is not yet possible to configure thrift server to use TLS.

Solution

Add protected ThriftServer#tlsEngine parameter, make it optional
(wrap () => Engine into Option] and set it to be None by default.
Use it to configure the server with Transport.TLSServerEngine instance
when it is overrided to be Some(x)

Result

The ability to use TLS with Thrift Server allowing secure
communications.