finatra-http: Ensure headers are set correctly
Problem

There are occurrences where we apply headers to the headerMap
using headerMap#add which will add an entry but not overwrite
an existing entry. This is generally incorrect for our usage where
the underlying assumption seems to be that the value will be
over written for the same key.

Solution

We should use headerMap#set where appropriate which will
overwrite a key/value in the map if it already exists.

Result

Correct application of headers in request and responses.

RB_ID=813969