finatra-thrift: Add ThriftWarmup for thrift servers.
Problem

There is currently no mechanism by which to issue warmup requests
through the filter and service chain defined by the ThriftRouter.

Solution

Implement a ThriftWarmup which exposes a way to send requests
to the implementing ThriftController which uses the filter
chain defined by the ThriftRouter for warming up the server.

RB_ID=820771