finatra: Fix Java support for callbacks returning a Future and add java-http-server to root build
Problem

The JavaController is written such that it's callback type
is always a java.lang.Object. The current CallbackConverter
does not properly deal with this causing the converter to
incorrectly wrapp the return in another Future.

The java-http-server example is not in the sbt root build as
it needed JDK 8 which we did not yet officially support.

Solution

Fix the CallbackConverter to deal with the return of the
JavaController correctly.

Now that we support JDK 8 we can add java-http-server example
into the sbt root build.

RB_ID=834467