Add sbt-revolver to the hello-world example
Problem

The documentation does not mention anything about reloading the server when
changes are made. There have also been issues created requesting this feature.

Solution

sbt-revolver is a plugin that provides the ability to reload(compile and run)
the server when source files are changed.

Result

The hello-world example can now be run with `sbt "~re-start"` to enable the
feature.