finagle-memcache: Add a boolean parameter to disable host ejection from KetamaFailureAccrualFilter
Problem

In HTTP, the Content-Length header is supposed to be the number of bytes (not
encoded characters) in the content body. There are 3 methods in
EmbeddedHttpServer (httpPut, httpPost, httpPatch) which are setting the
content length to the String length of the content, which is number of
characters.  With ASCII characters, those 2 lengths are the same. But if
the content contains any characters which will encode to more than 1 byte in
UTF-8, the string length will be smaller than the byte length, and the request
will essentially be corrupted.  E.g. a soccer ball emoji is \u26BD, a single
character in UTF-16, but when encoded into UTF-8, becomes 3 bytes -- E2 9A BD.
So if you set Content-Length to 1, the last 2 bytes will be unavailable to the
request handler.

Solution

When setting Content-Length, changed the 3 EmbeddedHttpServer methods (httpPut,
httpPost, httpPatch) to use the length of the UTF-8 encoded byte buffer instead
of the String length.

Result

Developers can successfully write tests using the httpPut, httpPost and
httpPatch methods of EmbeddedHttpServer, where the request body contains
characters whose UTF-8 encoding is more than 1 byte long.

RB_ID=836569