examples: Fixes Incorrect Example Documentation Links
Problem

There are several places in the examples have links that point
to incorrect locations resulting in 404 errors in some cases.

Solution

This pull request adds the correct links to those places in the documentation
as well as includes the "Full Changelog"
link that was missing in the `CHANGELOG.md` document.

Result

Working links in documentation/

Signed-off-by: Christopher Coco <ccoco@twitter.com>

RB_ID=838787