finatra: Add BUILD configuration for finatra/examples/hello-world-heroku
Problem

This project requires Scala 2.11+ as it uses a dependency only available
in Scala 2.11+ (com.github.rlazoti:finagle-metrics). As such it was not
configured for internal building as we supported Scala 2.10 and 2.11.

Solution

Since we no longer support Scala 2.10 we can add the proper BUILD
configuration for the project.

Result

Better internal visibility when refactoring.

RB_ID=842971