finatra-httpclient: Pass expected POST body to InMemoryHttpService map
Problem

The InMemoryHttpService allows tests to specify an expected body when mocking a
POST request.  This body is not being passed through to the map of responses, so
tests which should fail are passing.  Additionally, the test for this
functionality is faulty, because a generic "Exception" is expected to be thrown
when a non-expected body is received.  The assertFailedFuture test helper method
does not work for generic Exceptions, because the ScalaTest fail() method throws
an exception which is also caught.

Solution

Pass the "withBody" parameter from "mockPost" through to the responseMap.
Introduce PostRequestWithIncorrectBodyException so that the  "post mocked with
different body" test reports correctly.

Signed-off-by: Christopher Coco <ccoco@twitter.com>

RB_ID=844243