finatra: Narrow visibility on classes/objects in internal packages
Problem

In general, users of the framework SHOULD NOT rely on code in internal
packages, e.g., `com.twitter.finatra.http.internal`. These are meant
to be private to the framework and not intended as publicly accessible
APIs as they are implementation details specific to the framework and
are thus more subject to breaking changes. Unfortunately, we have been
lax in enforcing visbility on classes and objects in internal packages.

Solution

Narrow visibility as much as possible on classes and objects in
internal packages.

Result

Better intent of usage, e.g., these should not be used outside of the
framework. And if they are it should be a very conscious decision on
the part of the user (e.g., they have to put their scala code in a
package with the right visibility).

RB_ID=845278
TBR=true