finatra: Address lifecycle around com.twitter.inject.app.App#appMain
Problem

The function `appMain` in `com.twitter.inject.app.App` is confusingly named and lack of
clear documentation has led to many incorrect usages. The Finatra server lifecycle
(before this change) can be found [here](http://twitter.github.io/finatra/user-guide/getting-started/#lifecycle).

The important thing to note is that the Finatra framework defines and controls the `main`
method expected from a [TwitterServer](http://twitter.github.io/twitter-server/). This is so the framework can properly control
lifecycle events (which includes setting up the injector).

The goal of the Finatra framework is to make it easy to create a server without having to
worry about this lifecycle as it will be correctly handled by the framework.

In simple terms, the lifecycle can be diagramed like as follows. When you run:
java -Dfoo=bar -jar myapplication.jar -flag=value

`com.twitter.app.App#main(String[] args)`
  --> `com.twitter.app.App#nonExitingMain` (calls inits, parses flags, premains)
    --> `com.twitter.server.TwitterServer#main`
      --> `com.twitter.inject.server.TwitterServer#main`
      --> `com.twitter.inject.app.App#main` (calls finatra lifecycle methods)
        --> `com.twitter.inject.app.App#run`
        --> `com.twitter.inject.server.TwitterServer#start`
  --> (postmains, Await(this))
  --> execute all closes
  --> `com.twitter.app.App#main(String[] args)` exits

The original intent of the `appMain` method was as a final callback for users to implement
any additional server construction *or* to provide their application implementation. The
conflation of these two usages is somewhat confusing and hard to correctly handle.

The method `appMain` is run after all lifecycle methods have completed but *for servers*
(unlike a generic main method for [TwitterServer](http://twitter.github.io/twitter-server/)) it was **not expected** that users
block in this method as they will prevent the server lifecycle from correctly finishing.
However, for applications this is not true as this is where the main of the application is
expected to be written. However, this was not made clear and hard to reason about having
the same method name for both apps and servers.

Solution

This is a rather substantive change for the Finatra framework. It includes a lot of
changes resulting from trying to deal with the lifecycle around `appMain` which ended up
touching lots of parts of the codebase. We also add a lot of scaladoc documentation. When
in doubt, please check the scaladocs for more detailed information.

 Major changes:
- `com.twitter.inject.app.App#appMain` has been renamed. If you are
  using `com.twitter.inject.app.App` it is now, `run`, which is overriden to `start`
  in `com.twitter.inject.server.TwitterServer`. This can be read as you
  *run apps and start servers*.
- The `com.twitter.inject.app.EmbeddedApp` testing utility has been completely re-written
  to better deal with testing command line applications which is the primary intent for
  using `com.twitter.inject.app.App` directly. As such, this has affected
  `com.twitter.inject.server.EmbeddedTwitterServer`.
- The server framework previously tried to choose an Awaitable to await on. By default it
  would await on the [TwitterServer](http://twitter.github.io/twitter-server/) `adminHttpServer`.
  When implementing `com.twitter.finatra.http.HttpServer` it would instead chose the external
  httpServer and likewise when implementing `com.twitter.finatra.thrift.ThriftServer` the
  underlying finagle thriftServer. This has several consequences only some of which were
  alleviated due to current limitations in the Twitter util stack for Awaitables. When
  the framework chose to await on the external interface if the admin server interface exited
  the overall server process would not close. We also did not properly handle awaiting
  on https over http server. Additionally, if users blocked in the `appMain` method we never
  executed the server awaiting. Thus we've added a callback to register awaitables and we
  now await on them all in sequence. This is not ideal but `com.twitter.util.Await` currently
  lacks a fail-fast awaiting (like Future.select). We need to provide a API that allows
  Awaitables to produce a future that can be sequenced to onClose of the server.
- New lifecycle method, `postWarmupComplete` along with corresponding hook into the
  TwitterModuleLifecycle. This is for actions to be peformed after the injector has been
  created but before the server/application is marked as started.

RB_ID=844303
TBR=true