finatra: Aggregate inject/inject-thrift module in the sbt root
Problem

The inject/inject-thrift module is not aggregated in the
root build. Thus tests are not being run.

Solution

Add the inject/inject-thrift module to the list of
aggregated projects in the project root.

Result

Tests in the inject/inject-thrift module will be run and
code coverage should now be correct.

RB_ID=847601
TBR=true