finatra-thrift: Support dark traffic filtering
Problem

It is not currently possible to use the Finagle DarkTrafficFilter
in the finatra-thrift filter chain as the type signatures are
incompatible.

Solution

Introduce a Finatra-compatible DarkTrafficFilter (with a corresponding
module to provide it to the injector).

Result

It will be possible to send "dark" traffic to another service
while servicing requests in a Finatra thrift server.

RB_ID=852338