finatra-http: Add content headers for EmbeddedHttpServer #httpDelete and #httpPatch methods
Problem

The HTTP specification does not explicitly prohibit content bodies in HTTP DELETE requests and
the EmbeddedHttpServer#httpDeleteJson method (which uses #httpDelete) underneath does not
currently set the content-length or content-type headers in the request sent to the embedded
server. Similarly for #httpPatchJson (which uses #httpPatch), though here it is an oversight as this
method PATCH can be expected to have a content body (and thus a length) and type.

Solution

Add the content-type and content-length headers onto the request sent to the underlying embedded
server for EmbeddedHttpServer#httpPatch and  EmbeddedHttpServer#httpDelete.

Result

Proper requests for use in testing the embedded server.

RB_ID=862200