finatra-examples: Add web/UI open-source example
Problem

We do not have a simple open-source example for how to do basic file-serving
with Finatra. We field a lot of questions in this area and it would help to
have an example to which to refer users.

Solution

Add a basic example that shows how to return mustache views, files and a
"single-page" application.

RB_ID=868027