[AsyncStream] it consumes 34 secs for only one 104 KB string request
Problem

The endpoint with request type AsyncStream is time-consuming. From test, it
totally consumes 34 seconds in 38 seconds for only one string request with 104KB
(Note, 38 seconds is duration of http request completion). The root cause is
that the method buf.utf8str is time-consuming and called by
ByteBufferUtils.debugBuffer.

Solution

Since ByteBufferUtils.debugBuffer is for debugging purpose, we can limit it to
run only in debugging mode by logger.isDebugEnabled

Result

After applying solution, the time consumption is reduced to 0.254 seconds. It
improves a lot.