finatra: Move installation of the SLF4JBridgeHandler to the constructor of c.t.inject.server.TwitterServer
Problem

Currently the `Slf4jBridgeModule` installs the SLF4JBridgeHandler in the
`TwitterModuleLifecycle#singletonStartup` method. This method is run after
all premains from inside the `c.t.inject.app.App#main` in the
'postInjectorStartup' Finatra lifecycle method. This is rather late
and causes slf4j to miss some application/server initialization logging
since the bridges are not installed until after a fair amount of framework
initialization has happened. Further complicating the issue, is the fact
that the `c.t.logging.Logging` trait mixed into the
`c.t.server.TwitterServer` trait wipes out any added bridges
in it's premain, so even if we installed the bridges earlier they would
be wiped out in the premain of the application/server startup.

Solution

Add a new module: finatra/inject/inject-slf4j. This module provides
a utility for attempting to install the SLF4JBridgeHandler (when the
slf4j-jdk14 logging implementation is not on the classpath). Additionally,
it provides a `LoggerModule` which uses the utility to add
SLF4JBridgeHandler in the singletonStartup of the module lifecycle. Most
users should not need to worry about the `LoggerModule`, (neé
`Slf4jBridgeModule`) as it is only neccessary if you are not using the
`com.twitter.inject.server.TwitterServer` or a subclass, e.g., you are
building a command line application directly with
`com.twitter.inject.app.App`.

Move SLF4JBridgeHandler installation to the constructor of the
`c.t.inject.server.TwitterServer` as this will install the SLF4JBridgeHandler
at server instantiation. Also, with changes in the util-logging we can now
override the installation of the default logger factories (and handlers)
with an empty function so as not to remove any of added bridges. This allows
us to bridge util-logging implementation as early as possible to slf4j-api,
thus capturing more of the util-logging messages in slf4j.

Result

Users will be able to capture more of their application/server
initialization logging as we install the SLF4JBridgeHandler earlier.

RB_ID=870913