inject-thrift-client: Update filter building API with FilteredThriftClientModule
Problem

The `c.t.inject.thrift.filters.ThriftClientFilterChain` builder
API currently allows users to do bad things when constructing a
filter chain as the helper methods do not enforce a correct order
(leaving it to users to figure this out) and also allow repeated
invocations to append the created filters to the filter chain
multiple times. Additonally, we'd like a helper for installing a
`LatencyFilter` which can track the latency of both a method
invocation (e.g., including retries) and per request made (e.g,
each retry).

Solution

We've updated the API for consistency with other builder APIs in
the codebase. Functionally, the changes should be backwards
compatible only now we make sure that each helper method now
installs and configures a filter specific to their function, i.e.,
either a RetryFilter, a LatencyFilter, a Filter for exception
handling, or a TimeoutFilter, etc. These filters are then composed
in a correct order on top of the general filter chain which is
built up by calls to `#filtered` or `withAgnosticFilter`.

Note that we've also decoupled the installation of either a constant
or exponential retry filter with adding a request timeout filter.
While, well-intentioned (you should almost always specify a request
timeout in some fashion) it is awkward to link these concerns in
the filter chain builder API.

Result

Better semantics and utility around building a Thrift client which
filters per-method.

RB_ID=878260
TBR=true