finatra-http ResponseBuilder.fileOrIndex verify when file exists
Problem

ResponseBuilder.fileOrIndex method is not verifying when the file exists in
the file system but validating that it contains a file extension. This leads
to get http error 404 (Not found) when the file has an extension and doesn't
exists (see issue #360).

Solution

Lets implement a method hasFile at ResponseBuilder which checks when the
file exists and can be retrieved by fileResolver.getInputStream method.

Modify method fileOrIndex from ResponseBuilder to check when the file exists
to avoid http error 404.

Result

Calling method fileOrIndex from ResponseBuilder won't return http error 404
when file doesn't exists but return index.